pyproject.toml
[project]
name = "pookie-audio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.6.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.28.0",
]
